+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Simple load balancing".
 - Tests for "Are the clumps normal".
 - Tests for "Bulk email generator".
+- Tests for "Snail run".
 
 ## [3.8.0] - 2022-04-30
 ### Added